refactor: 해당 entity 일관성 있게 리팩토링
