+++ src/main/java/com/project/trainingdiary/entity/BlacklistedTokenEntity.java
@@ -2,15 +2,13 @@ package com.project.trainingdiary.entity;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.Id;
-import jakarta.persistence.Table;
 import java.time.LocalDateTime;
 import lombok.Getter;
 import lombok.Setter;
 
-@Entity
 @Getter
 @Setter
-@Table(name = "blacklisted_token")
+@Entity(name = "blacklisted_token")
 public class BlacklistedTokenEntity {
 
   @Id
