+++ src/main/java/com/project/trainingdiary/entity/VerificationEntity.java
@@ -2,7 +2,6 @@ package com.project.trainingdiary.entity;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.Id;
-import jakarta.persistence.Table;
 import java.time.LocalDateTime;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -10,13 +9,12 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
-@Entity
-@Table(name = "verification")
 @Getter
 @Setter
 @AllArgsConstructor
 @NoArgsConstructor
 @Builder
+@Entity(name = "verification")
 public class VerificationEntity extends BaseEntity {
 
   @Id
